[Model] Address review: GVA support, A-state gradients, OOB fixes
Addresses the review round on the precond ops:

1. Implement GVA (num_v_heads > num_heads) properly for PGDN instead of
   computing garbage: the KKT kernel now carries HV heads for g/beta/A with
   i_hv // (HV // H) key-head indexing (the wy bwd path already supported it),
   and the ATK preconditioner is pinned to the key-head (H) contract
   throughout - layer projections/params, fused_recurrent indexing, and the
   A-state shape [N, H, K] - matching what the docstrings promised. The naive
   reference applies the GVA repeat internally, and new test_chunk_gva cases
   verify fwd + all grads against it at 2x and 4x group sizes.

2. Propagate the final A-state gradient: backward no longer drops dat. Since
   at == ac[:, last_chunk], the incoming gradient is injected into the ATK
   reverse-scan carry at the final chunk and flows to dk/dbeta_atk/dg_atk/
   dh0_atk. New test_chunk_A_state cases (both ops) pass a nonzero
   initial_A_state, keep at in the loss, and check at/dh0_atk against
   autograd through the naive reference.

3. Raise NotImplementedError for cp_context instead of silently computing
   wrong results; docstrings updated accordingly.

4. Fix out-of-bounds loads: the unmasked g load in precond_kda
   fused_recurrent (same bug #1135 fixed in baseline KDA) plus unmasked
   gk/gv loads in the PGDN fused_recurrent.

5. Adapt LegacyFLACache.update for the A_state slot so generation on older
   transformers no longer raises TypeError.

Also in this round: restore int64 index guards across the precond/ATK
kernels (baseline program_id idiom), register both ops in
benchmarks/ops/registry.py, and declare a_log_init_lower as a documented
config knob instead of a hidden getattr.

Full precond suite on A100: 85 op tests + 7 model tests pass, 0 failed;
GVA model checks (PGDN + PKDA) finite; CP raise verified.

+++ fla/ops/atk/chunk_atk_fwd.py
@@ -54,17 +54,17 @@ def _forward_chunk_summary(
 ):
     i_t = tl.program_id(0)
     h = tl.program_id(1)
-    chunk_id = tl.program_id(2)
+    chunk_id = tl.program_id(2).to(tl.int64)
 
     if IS_VARLEN:
         i_n = tl.load(chunk_indices + i_t * 2).to(tl.int32)
-        chunk_id = tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
+        chunk_id = tl.load(chunk_indices + i_t * 2 + 1).to(tl.int64)
         bos = tl.load(cu_seqlens + i_n).to(tl.int32)
         eos = tl.load(cu_seqlens + i_n + 1).to(tl.int32)
         T = eos - bos
-        b = i_n
+        b = i_n.to(tl.int64)
     else:
-        b = tl.program_id(0)
+        b = tl.program_id(0).to(tl.int64)
         bos = 0
         eos = T
 
@@ -136,7 +136,7 @@ def _forward_pass_chunks(
     USE_INITIAL_STATE: tl.constexpr,
     IS_VARLEN: tl.constexpr,
 ):
-    i_nh = tl.program_id(0)
+    i_nh = tl.program_id(0).to(tl.int64)
 
     if IS_VARLEN:
         i_n = i_nh // H
@@ -144,7 +144,7 @@ def _forward_pass_chunks(
         bos = tl.load(cu_seqlens + i_n).to(tl.int32)
         eos = tl.load(cu_seqlens + i_n + 1).to(tl.int32)
         T = eos - bos
-        b = i_n
+        b = i_n.to(tl.int64)
     else:
         b = i_nh // H
         h = i_nh % H
@@ -216,17 +216,17 @@ def _forward_chunk_out(
 ):
     i_t = tl.program_id(0)
     h = tl.program_id(1)
-    chunk_id = tl.program_id(2)
+    chunk_id = tl.program_id(2).to(tl.int64)
 
     if IS_VARLEN:
         i_n = tl.load(chunk_indices + i_t * 2).to(tl.int32)
-        chunk_id = tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
+        chunk_id = tl.load(chunk_indices + i_t * 2 + 1).to(tl.int64)
         bos = tl.load(cu_seqlens + i_n).to(tl.int32)
         eos = tl.load(cu_seqlens + i_n + 1).to(tl.int32)
         T = eos - bos
-        b = i_n
+        b = i_n.to(tl.int64)
     else:
-        b = tl.program_id(0)
+        b = tl.program_id(0).to(tl.int64)
         bos = 0
         eos = T
