+++ fla/ops/precond_kda/chunk_intra_token_parallel.py
@@ -50,7 +50,7 @@ def chunk_precond_kda_fwd_kernel_intra_token_parallel(
     - Aqk = q @ k_precond^T (column uses k_precond)
     - Akk = k @ k_precond^T (row uses original k, column uses k_precond)
     """
-    i_tg, i_hg = tl.program_id(0), tl.program_id(1)
+    i_tg, i_hg = tl.program_id(0).to(tl.int64), tl.program_id(1)
 
     if IS_VARLEN:
         i_n = 0
@@ -68,7 +68,7 @@ def chunk_precond_kda_fwd_kernel_intra_token_parallel(
                     left = mid + 1
         i_n = left
 
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int32), tl.load(cu_seqlens + i_n + 1).to(tl.int32)
+        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
         T = eos - bos
         i_t = i_tg - bos
     else:
